structure with 4 crates
