+++ crates/commands_robbb/Cargo.toml
@@ -13,14 +13,12 @@ chrono-humanize = "0.2"
 itertools = "0.10"
 indoc = "1.0"
 lazy_static = "1.4"
-rand = "0.8"
 humantime = "2"
 thiserror = "1.0"
 serde_json = "1"
 serde = "1"
 maplit = "1"
 byte-unit = "4.0"
-url = "2"
 regex = "1"
 reqwest = { version = "0.11" }
 tokio = { version = "1.16", features = ["macros", "fs", "rt-multi-thread"]}
@@ -30,9 +28,7 @@ futures = "0.3"
 unicase = "2.6.0"
 
 tracing = "0.1.31"
-tracing-log = "0.1.2"
 tracing-futures = "0.2.5"
-tracing-subscriber = { version = "0.3.0", features = [ "std", "env-filter" ] }
 poise = "0.1.0"
 parking_lot = "0.12.0"
