+++ crates/commands_robbb/src/commands/fetch/fetch.rs
@@ -1,4 +1,4 @@
-use shared_robbb::fetch_field::FetchField;
+use db_robbb::fetch_field::FetchField;
 
 use super::*;
