+++ crates/commands_robbb/src/commands/tag.rs
@@ -1,4 +1,5 @@
 use chrono::Utc;
+use shared_robbb::prelude::{AppCtx, Ctx, Res};
 
 use super::*;
 
@@ -134,7 +135,7 @@ pub async fn tag_set(
 
     let existing_tag = db.get_tag(&tag_name).await?;
 
-    let new_content = util::run_text_field_modal(
+    let new_content = run_text_field_modal(
         app_ctx,
         "Set your tag",
         "Content",
@@ -153,4 +154,59 @@ pub async fn tag_set(
     .await?;
     ctx.say_success("Succesfully set!").await?;
     Ok(())
+}
+
+pub async fn run_text_field_modal(
+    app_ctx: AppCtx<'_>,
+    title: &str,
+    label: &str,
+    placeholder: &str,
+    default_content: &str,
+) -> Res<String> {
+    // For now we must do this manually rather than using the poise macro solution
+    // as we want to set a default value for the modal content.
+
+    let interaction = app_ctx.interaction.unwrap();
+    interaction
+        .create_interaction_response(app_ctx.discord, |ir| {
+            ir.kind(poise::serenity_prelude::InteractionResponseType::Modal);
+            ir.interaction_response_data(|d| {
+                d.custom_id("text_field_modal");
+                d.title(title);
+                d.components(|c| {
+                    c.create_action_row(|r| {
+                        r.create_input_text(|t| {
+                            t.custom_id("content");
+                            t.label(label);
+                            t.style(poise::serenity_prelude::InputTextStyle::Paragraph);
+                            t.required(true);
+                            t.value(default_content);
+                            t.placeholder(placeholder)
+                        })
+                    })
+                })
+            })
+        })
+        .await?;
+
+    app_ctx
+        .has_sent_initial_response
+        .store(true, std::sync::atomic::Ordering::SeqCst);
+
+    // Wait for user to submit
+    let response = poise::serenity_prelude::CollectModalInteraction::new(&app_ctx.discord.shard)
+        .author_id(interaction.user.id)
+        .await
+        .unwrap();
+
+    // Send acknowledgement so that the pop-up is closed
+    response
+        .create_interaction_response(app_ctx.discord, |b| {
+            b.kind(poise::serenity_prelude::InteractionResponseType::DeferredUpdateMessage)
+        })
+        .await?;
+
+    let content = poise::find_modal_text(&mut response.data.clone(), "content")
+        .user_error("Missing tag content data from modal")?;
+    Ok(content)
 }
