+++ crates/db_robbb/Cargo.toml
@@ -7,17 +7,15 @@ edition = "2021"
 
 [dependencies]
 serenity = { version = "0.11.1", default-features = false, features = ["collector", "builder", "cache", "chrono", "client", "gateway", "model", "http", "utils"]}
+poise = "0.1.0"
+thiserror = "1.0"
 sqlx = { version ="0.5", features = ["runtime-tokio-rustls", "sqlite", "chrono", "offline"] }
 anyhow = "1.0"
 chrono = "0.4"
-chrono-humanize = "0.2"
 itertools = "0.10"
 regex = "1"
 tracing = "0.1.31"
-tracing-futures = "0.2.5"
 serde_json = "1"
 serde = "1"
 lazy_static = "1.4"
 unicase = "2.6.0"
-
-shared_robbb = { path = "../shared_robbb" }
