+++ crates/db_robbb/src/db/fetch.rs
@@ -5,8 +5,8 @@ use anyhow::{Context, Result};
 use chrono::{DateTime, Utc};
 use serenity::model::id::UserId;
 
+use super::fetch_field::{FetchField, FETCH_KEY_ORDER};
 use super::Db;
-use shared_robbb::fetch_field::{FetchField, FETCH_KEY_ORDER};
 
 #[derive(Debug)]
 pub struct Fetch {
