+++ crates/db_robbb/src/db/mod.rs
@@ -12,6 +12,7 @@ pub mod ban;
 pub mod blocklist;
 pub mod emoji_logging;
 pub mod fetch;
+pub mod fetch_field;
 pub mod highlights;
 pub mod kick;
 pub mod mute;
