+++ crates/robbb/src/events/guild_member_addition.rs
@@ -1,5 +1,7 @@
 use super::*;
 use chrono::{DateTime, Utc};
+use commands_robbb::log_error;
+use shared_robbb::util;
 use std::time::SystemTime;
 
 // TODORW this belongs in crate::commands::mute
