+++ crates/shared_robbb/src/embeds.rs
