+++ crates/shared_robbb/src/extensions.rs
